Make batch_size uint64_t instead of size_t

+++ tg_api/silkworm_tg_api.cpp
@@ -24,7 +24,7 @@
 #include <silkworm/execution/execution.hpp>
 
 SILKWORM_EXPORT SilkwormStatusCode silkworm_execute_blocks(MDB_txn* mdb_txn, uint64_t chain_id, uint64_t start_block,
-                                                           size_t batch_size, bool write_receipts,
+                                                           uint64_t batch_size, bool write_receipts,
                                                            uint64_t* last_executed_block,
                                                            int* lmdb_error_code) SILKWORM_NOEXCEPT {
     assert(mdb_txn);
