+++ tg_api/silkworm_tg_api.h
@@ -70,7 +70,7 @@ enum SilkwormStatusCode {
  * (blocks up to and incl. last_executed_block were still executed).
  */
 SILKWORM_EXPORT SilkwormStatusCode silkworm_execute_blocks(MDB_txn* txn, uint64_t chain_id, uint64_t start_block,
-                                                           size_t batch_size, bool write_receipts,
+                                                           uint64_t batch_size, bool write_receipts,
                                                            uint64_t* last_executed_block,
                                                            int* lmdb_error_code) SILKWORM_NOEXCEPT;
